shill: don't truncate sequence numbers
Both when handling NetlinkPackets and when managing the pending dump
sequence number shill has previously truncated the 32bit netlink
sequence to a uint16_t.  This causes issues where the untruncated
sequence is used as the key in the map of handlers, meaning that the
handler never gets invoked and the messages are then treated as if they
were broadcast messages.

Bug: None
BUG=chrome-os-partner:46556
TEST=created a program that issued more than 2^16-1 dump requests

+++ shill/net/netlink_manager.cc
@@ -617,7 +617,7 @@ bool NetlinkManager::IsDumpPending() {
   return dump_pending_ && !pending_messages_.empty();
 }
 
-uint16_t NetlinkManager::PendingDumpSequenceNumber() {
+uint32_t NetlinkManager::PendingDumpSequenceNumber() {
   if (!IsDumpPending()) {
     LOG(ERROR) << __func__ << ": no pending dump";
     return 0;
